Osa4 blog 4.15 -4.23

+++ controllers/Blogs.js
@@ -1,33 +1,82 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/Blog')
+const User = require("../models/User")
+const jwt = require("jsonwebtoken")
 
+
+const getTokenFrom = request => {
+  const authorization = request.get('authorization')
+  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+    return authorization.substring(7)
+  }
+  return null
+}
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate("user", {
+      username: 1,
+      name: 1,
+      id: 1,
+    })
     response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response) => {
   
   const { title, author, url, likes } = request.body
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  console.log(decodedToken.id)
+  
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({
+      error: 'token missing or invalid'
+    })
+  }
+
+  const user = await User.findById(decodedToken.id)
 
   const blog = new Blog({
     title,
     author,
     url,
-    likes: likes || 0
+    likes: likes || 0,
+    user: user._id
   })
-  await blog.save()
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
   response.status(201).json(blog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
 
-  await Blog.findByIdAndRemove(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({
+      error: 'token missing or invalid'
+    })
+  }
+
+const blog = await Blog.findById(request.params.id)
+
+  if (!blog.user || blog.user.toString() === request.user._id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+  } 
+  else {
+    response.status(403).end()
+  }
   response.status(204).end()
 })
 
 blogsRouter.put('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({
+      error: 'token missing or invalid'
+    })
+  }
 
   const blog = {
     title: request.body.title,
@@ -36,11 +85,8 @@ blogsRouter.put('/:id', async (request, response) => {
     likes: request.body.likes
   }
 
-  const updatedBlog = await Blog.findByIdAndUpdate(
-    request.params.id,
-    blog,
-    { new: true }
-  )
+  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, 
+    { new: true }).populate('user', { username: 1, id: 1, name: 1 })
 
   response.json(updatedBlog)
 })
